+++ docs/changelog.md
@@ -18,6 +18,7 @@ category: Administration
 * chore: rename Git master branch to main
 * feat: support latest enterprise patch releases for Jira 7.13, 8.5 and 8.13
 * refactor: cleanup code for health check action
+* test: add tests for Scrum Poker session action
 
 ### [20.11.0] - 2020-11-14